refactor to avoid Clerk code in local auth mode
